+++ Ryn.Windowing/Include/Ryn/Windowing/Window.hpp
@@ -2,7 +2,7 @@
 
 #include <Ryn/Core/Utility.hpp>
 #include <Ryn/Core/Function.hpp>
-#include <Ryn/Numerics/Vector2.hpp>
+#include <Ryn/Mathematics/Vector2.hpp>
 #include <Ryn/Windowing/Events/Event.hpp>
 #include <Ryn/Windowing/WindowSettings.hpp>
 
@@ -22,11 +22,11 @@ namespace Ryn
 
         virtual void Close() = 0;
 
-        virtual void Resize(Vector2<u16> size) = 0;
+        virtual void Resize(Vector2<u32> size) = 0;
         virtual void SetTitle(const String& title) = 0;
 
-        virtual Vector2<u16> GetSize() const = 0;
-        virtual Vector2<u16> GetFramebufferSize() const = 0;
+        virtual Vector2<u32> GetSize() const = 0;
+        virtual Vector2<u32> GetFramebufferSize() const = 0;
         virtual String GetTitle() const = 0;
 
         void SetEventCallback(Function<void(Event&)>&& callback) { EventCallback = Utility::Move(callback); }
