+++ Ryn.Windowing/Source/Platform/Cocoa/CocoaWindow.hh
@@ -18,11 +18,11 @@ namespace Ryn
 
         virtual void Close() override;
 
-        virtual void Resize(Vector2<u16> size) override;
+        virtual void Resize(Vector2<u32> size) override;
         virtual void SetTitle(const String& title) override;
 
-        virtual Vector2<u16> GetSize() const override;
-        virtual Vector2<u16> GetFramebufferSize() const override;
+        virtual Vector2<u32> GetSize() const override;
+        virtual Vector2<u32> GetFramebufferSize() const override;
         virtual String GetTitle() const override;
 
         void OnEvent(Event& event) { if (EventCallback) EventCallback(event); }
